Move the UserDetailsService to a new package for services
