+++ src/main/java/org/apereo/portlet/soffit/renderer/SoffitRendererController.java
@@ -11,10 +11,10 @@ import javax.servlet.http.HttpServletResponse;
 import org.apache.commons.lang3.StringUtils;
 import org.apereo.portlet.soffit.Headers;
 import org.apereo.portlet.soffit.connector.SoffitConnectorController;
-import org.apereo.portlet.soffit.connector.UserDetailsService;
 import org.apereo.portlet.soffit.model.v1_0.Payload;
 import org.apereo.portlet.soffit.model.v1_0.Request;
 import org.apereo.portlet.soffit.model.v1_0.UserDetails;
+import org.apereo.portlet.soffit.service.UserDetailsService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
