+++ src/main/resources/org/apereo/portlet/soffit/mvc/soffit-connector.xml
@@ -19,6 +19,8 @@
     <!-- Allows the SoffitConnectorController to benefit from uPortal's PlaceholderConfigurer -->
     <bean parent="primaryPropertyPlaceholderConfigurer" />
 
+    <bean class="org.apereo.portlet.soffit.service.UserDetailsService" />
+
     <!-- This is a bit of a mess:  the Java class is a part of uPortal (not Soffit), but it's referenced in this file.  When this tech is moved into the uPortal project, this mismatch will go away. -->
     <bean class="org.jasig.portal.soffit.AuthorizationHeaderProvider" />
